+++ feature/dailyHugg/src/main/java/com/hugg/dailyhugg/all/DailyHuggListEvent.kt
@@ -1,0 +1,4 @@
+package com.hugg.dailyhugg.all
+
+class DailyHuggByRoundEvent {
+}
