+++ feature/dailyHugg/src/main/java/com/hugg/dailyhugg/all/DailyHuggListPageState.kt
@@ -1,0 +1,7 @@
+package com.hugg.dailyhugg.all
+
+import com.hugg.feature.base.PageState
+
+data class DailyHuggByRoundPageState(
+
+): PageState
