--- feature/dailyHugg/src/main/java/com/hugg/dailyhugg/byRound/DailyHuggByRoundEvent.kt
@@ -1,4 +1,0 @@
-package com.hugg.dailyhugg.byRound
-
-class DailyHuggByRoundEvent {
-}
