--- feature/dailyHugg/src/main/java/com/hugg/dailyhugg/byRound/DailyHuggByRoundPageState.kt
@@ -1,4 +1,0 @@
-package com.hugg.dailyhugg.byRound
-
-class DailyHuggByRoundPageState {
-}
